cdk/pyproject.toml
[tool.poetry]
name = "cardiff-iac"
version = "0.0.1"
description = "Infrastructure as code for Cardiff Twitter Bot"
authors = ["Michael Brewer"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.33.0"
constructs = "^10.1.57"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
